requirements.txt
# Betfair Exchange Trading Bot - Dependencies

# Betfair API
betfairlightweight>=2.20.0

# Web framework and async
aiohttp>=3.9.0
requests>=2.31.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0  # PostgreSQL async driver

# Scheduling
apscheduler>=3.10.0

# Telegram bot
python-telegram-bot>=20.7

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# Logging and monitoring
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
ruff>=0.1.0

# HTTP retry logic
tenacity>=8.2.0
